requirements.txt
django==6.0
strawberry-graphql-django==0.70.1
django-choices-field==3.1.1
django-environ==0.12.0
pkg-filters[all] @ https://github.com/itqadem-apps/pkg_filters/releases/download/pkg_filters-v1.12.0/pkg_filters-1.12.0-py3-none-any.whl
pkg-auth[all] @ https://github.com/itqadem-apps/pkg_auth/releases/download/pkg_auth-v0.15.0/pkg_auth-0.15.0-py3-none-any.whl